Use Boost.HOF to implement the conditional matching logic behind transform().
Wrap the first argument to transform() by calling as_expr() on it, so that
as_expr() is rarely needed outside of transform().

Continue to accept rvalue transform objects in the top-level transform(), but
do not perfectly forward it into the implementation details of transform().
This fixes the potential bug of stealing the movable contents of a transform.

Fixes #62.